Several changes required to run on linux.
1. Makefile gets itself as a dependency.

2. Rearrange make.

3. Don't use the local version of stdio.h in pa_ansi.

4. Change to selects() to avoid system call by same name.

5. Our stdio needs to use its own header in front of others.

So what's going on here? Petit Ami needs to insert it's own
stdio ahead of the library set of libc. Libc has a bunch of
modules that are cross depdent on each other. What are doing
is like trying to change out one of the cards from a deck
that a dealer is trying to deal a hand from, without their
knowing about it. For this we rely on the fact that the
gcc linker will defer to local implemented symbols. Thus, we
let everyone link at compile time to the standard stdio.h
except stdio.c, our local trojan horse of a module. We can't
let our stdio implementation use the general stdio.h because
we have special fields in the FILE type.