1. Various updates.
2. Coined using "pa_" coin.

I decided to use a coin for all Petit_ami modules that this would unify
the entire PA set, and also to prevent the need to change the names when
advancing in module types.

This works best for C, but not for a name managed system like Pascaline
I think.