General refactor to add split buffer ability to ansiterm.c. Split buffer was added to later libraries such as console and graphics modes on windows. At that point, ansilib (old name for ansiterm) was rendered redundant by console mode. So we needed to update ansiterm.c to implement the new behavior.
1. Added indisp() function to check given screen is in display.

2. Across the board, where screens[curupd-1] was used to indicate the destination
or working buffer, I changed the routines to accept an input parameter for
the working screen. This is both more flexible and results in cleaner and more
compact code, since we are not recalculating the screen address all of the time.

3. Setcur implements cursor movement optimizations to check the current and target
cursor positions and then output the minimum ANSI sequence to get there. This was
necessary to fit setcur's new status as the center of cursor movements. Now we
divide the virtual screen state from the physical state, and motions like pa_up()
just change the virtual position, and setcur() and friends compare the virtual
and physical states and make them match. From now on, we refer to screen(s) as
the virtual screens, and display as the physical display.

4. iniscn() now just does buffer init, not display.

5. restore() updates physical position, since it always operates on the display.

6. iscroll() now considers if it is in display, and operates or not on the
physical display if so. Otherwise, all actions are buffer only.

7. iclear() is now virtual only.

8. iup(), idown(), ileft() and iright() now manipulate the virtual position of
the current update screen only, and then leave it to setcur() to sync the
display if the current screen is in display.

9. plcchr() considers if the screen is in display. Also, it uses the logic
in iright() do caluculate the actions after a character is placed. We needed to
do that because iright() updates the cursor, and in plcchr() that is a side
effect of the character placement itself. I'll consider better options later
(like a common routine).

10. pa_home() updates the phy position now.

11. pa_del() uses the i (internal) routines now. It's just cleaner.

12. pa_blink(), pa_reverse(), pa_underline(), pa_superscript(), pa_subscript(),
pa_italic(), pa_bold(), pa_strikeout(), pa_standout(), pa_fcolor() and pa_bcolor()
are changed to only update the display if in display.

Note the attribute sets all also set the colors, I think this was to fix the
reverse video attribute issue with ansi terminals previously mentioned, and can
be removed. Also, I think a setcolor() routine to sync the virtual to physical
colors would be appropriate.

13. pa_select() is filled out completely. It checks for any changes in update or
display screens, and if so, and the new location is not allocated, it allocates
a new screen and initalizes it. Then the display screen, if changed, is restored
from the new display screen.

14. The init block changes to use restore() to update the newly created screen.