1. Checked off add resize event.
2. Removed tear-off parameters. I researched it, it won't work except
in a extremely implementation dependent way. Bottom line is there is no
LIBC support for it at all, and that can't be gotten round like we did
with stdio. Its too bad, that capability should be in LIBC/GLIBC.

The alternative would have been a special call to init the PA arguments
from main(). That is still a possibility, but kinda negates the basic
idea of PA, which is the source does not need to be modified.

Going forward we will rely on the .termrc files.