Added note to fix snake game.
Coulda sworn I got rid of this before.

Modified the way the resize event works, now it tries to compensate the
display as well.

1. Created new routine, findsize(), and call it from the resize event.

2. On resize event, inpevt() saves the old size, gets the new size, and
checks if we are going to expose any new areas. If so, we do a redraw.

3. clrbuf() now initalizes the entire screen store, not just the
display area. Thus new screen buffer will do the right thing when
exposed.

4. In plcchr(), we don't clip the buffer writes when cursor out of
bounds anymore. Thus the buffer content will be there if exposed.

5. Remove size find from init and replace with findsize() call.