1. Changed test.c formula.
2. Fixed make clean.

Checked off getc() task.

1. Coupled getc() and putc() definitions to USEMACRO flag.

2. Removed diagnostics.

1. Added new switch, USEMACRO, that enables use of macros for getc() and
putc(). This is for any compatability issues with using hard functions
for these instead of macros.

2. Added definitions for getc() and putc(), in either macro or header
forms.

Changed include to directly use stdio.h. There are two such files, one
local to PA and the system definition. Unfortunately they cannot be
identical, and the compiler will flag them as being based on different
structs. The answer is to always use the PA stdio.h file. In the final
form of PA, stdio.h will simply be first in the include path.