Workaround for SIGWINCH resizing. Exlaination below.
Always performs a screen restore on SIGWINCH.

This works, but generates a lot of tearing and display flashes.

I experimented and observered a few things:

1. vi/curses does not have the issue. I put the same test screen
through those and resized them, no issue.

2. xterm/gnome wrap or don't wrap text they have buffered. You can
even turn the wrap mode off (printf("\33[?7l");) and it will clip, and not
wrap, new text but old text buffered on screen will be wrapped. I determined
this by printing (cat) a fixed file, with wrap off it will be displayed
in clipped mode, but even those clipped elements in the buffer will be
wrapped.

What does this matter. Well, the wrap is what is causing the display tearing.
Even spaces get wrapped, so varying the geometry of the window even causes
the line spacing to change.

I have two theories why that does not affect vi. First is they are using
a wrap-off procedure I am not aware of, and second turning off wrap
completely is a byproduct of alternate buffer flipping. We will see.

For now this works and is testable.