An issue was found during GTK adoption. We used a priority labeled constructor to init the stdio.c module. Unfortunately its possible to run code before all constructors execute by using .init section code, its the older method to do inits. Even if we also use that init method, we would then have to modify the link map to make it work out.
An easier solution was to move all the required initialization to compile
time or key it to NULL or zero value, meaning that an invalid 0 value kicks
off what would have normally occurred at init time. This means that it does
not matter when stdio routines are called, they will tolerate that.

1. Created static structures for each of the standard in, standard out and
standard error control blocks.

2. stdin, stdout and stderr are predefined to point to their control blocks.

3. The files table is preinit to point to the same definitions, which
emulates the previous code. Technically this is no longer really necessary.

4. The powers table, used for radix conversion, is created dynamically at
the spot it is used.

5. The vectored override calls are preinitialized.

6. Several undefined functions were filled out in efforts to reduce hitting
those with GTK calls (more on that below). These are fseek(), ftell(),
rewind(), fgetpos(), fsetpos(), and perror(). Note several are still
left undefined.

7. The init constructor is removed.

When GTK is pulled in, it pulls in a large amount of secondary libraries,
ones it uses or are used by those, etc. Thus stdio.c needs to step up it's
compatability to play at this level.

GTK still has bringup issues at this time. I am going to try some diagnostics
to see what is being called and failing. I verified via previous experiment
that the stack is indeed calling stdio.