Filled out the function catalog in stdio.c a bit more.
1. Added flags/status word. The append and text fields could be rolled
into this.

2. Added counters and local string storage for temp file names.

3. Changed some dangling routines to static.

4. Filled out functions: tmpfile(), tmpnam(), setvbuf(), setbuf(),
clearerr(), feof() and ferror().

5. On various routines, the eof status flag was set so that feof()
can check it. Its a bit redundant with the normal EOF check error.
We set it on any read that detects EOF, and clear on any seek.
Need to check if EOF is true on write files.

6. Added destructor to remove temp files. To some extent, there
is abiguity similar to the constructor situation, since another module
could pick the same priority. This is, IMHO, a fundamental defect
with the constructor/destructor system, which (for example)
Pascaline does not use. I suspect that the equivalent for the
Pascaline system is to use .init blocks and specifically arrange
the link time arrangements. See next.

Ok, since I mentioned it, how does Pascaline avoid the init issues?
Its because Pascaline modules stack, that is, each module has a specific
init and deinit handling at the front of the module, and then each
module chains the next in sequence. This allows the init/deinit
order to be specified exactly.