Added sun glibc extensions. These were being called by other programs, and they directly access FILE* formats. Since this was being handled by glibc, this was causing segfaults.
1. Support "e" mode flag, open across execs, a glibc extension.

2. Added some default inits, not technically required.

3. Added last read and last write flags to support new functions.

4. Added ext functions __fbufsize(), __fpending(), __flbf(), __freadable(),
__fwritable(), __freading(), __fwriting(), __fsetlocking(), __flushlbf(),
__fpurge().

The ext functions were put in the same stdio.c file, but the stdio_ext.h
file is used as per glibc. This is because programs use the header file,
but it does not matter where the actual functions are. glibc typically
uses one function per file (like printf.c).