End of experiment, extended stdio calls removed, see below.
1. Removed _EFRDL and _EFWRL flags and their associated code.

2. Removed all stdio_ext routines.

Removed _EFRDL and _EFWRL flags.

The inclusion of the ext routines fixed the crash issues with
selinux, but then it just fell over again with a new issue, libthreads_db,
which not only uses new internal glibc routines, but also paws through
the FILE* data structure as well.

I have been through the experience of trying to emulate other people's
software before. It's a never ending job. There is always the next bit
of code that is more intrusive, more intimately tied to the package,
and the target you are trying to emulate is changing as well. Its
not worth it.

The stdio "single plank" used here is a good tool, and works for a lot
of applications. It came from a full libc implementation I did for
embedded use. I am already aware that it does not work for certain
applications. For example using static linking does not work, since
my replacing stdio alone (without replacing the rest of libc) only
works when using dynamic linking (.so).

The answer for these applications is to use a modified glibc release
that supports hooking vectors to the read, write and other fundamental
functions. This fixes both the static linking issue and the backdoor
issues I have seen (since glibc emulates glibc quite well, thank you).

As a side comment, this explains why glibc is so hard to replace
in common linux applications, and also one reason why it is so
bloated. It is woven into the linux infrastructure like an octopus,
with tenacles everywhere. It doesen't even mean its glibc's fault.
All of these programs reaching into the internal workings of glibc.
Is this good programming?