Added a few utilities, a starting point for an X Windows graphical module.
Disables alternate buffer from the command line.

Enables alternate buffer from the command line.

Alternate buffer demo program.

Compile script for xtest.

Test skeleton for X Windows implementation of graphics module.

This is "hello world" for x from the internet.

After the fiasco with the GTK+ implementation stack, I created a
test for feasability with a graphics module that goes direct to
XWindows, which works quite well. This is a fallback/alternate plan
if the GTK+ system can't be made to work.