This commit is the override version of glibc. I am using a tip/master for glibc here. I will distribute the changes required to glibc only in the form of a patch here. The method to recreate is to get a copy of glibc source, apply the patch, and compile it. I have supplied the libc.a (static) and libc.so (dynamic) glibc libraries so you don't have to do this.
Refactored the make file to take options for both the internal stdio and
glibc override libraries. The change is incomplete, only test and scntst
are so converted.

This contains the override patch code. I don't have a correct unlink()
method at this point. The patch code does not appear to compile, I can put
an error in it and it makes fine. The issue is the over_unlink() call does
not get exposed as a global. Still working on that.

These are the static and dynamic override versions of glibc.

1. Left the stubouts in by accident, will fix next patch.

2. Removed if'ed out code to do speed tests. This relied on floating point
code, which stdio local version didn't have. Now glibc has this.

3. Character speed test was not being calculated in floating point, fixed
with cast.

4. Scrolling speed test was too long, cut down by 1/10th. I suspect this
means the Windows version was faster because it directly manipulates the
pixel buffer.

5. Buffer flip speed test was too long, cut down by 1/10th.

Needed to use stdio.h include from library path, not direct.

1. Turned off I/O buffering on stdin/stdout. I am not actually sure this
works.

2. Commented out ovr_unink() calls because of issues with that (see above).
We don't use this in any case.

3. Moved the override vector holding vars to the top of the routine. I suspect
gcc was allowing this using c++ rules, which is not right. We are supposed
to be ANSI C compliant.

4. Added fflush() calls to terminal mode changes outside of the raw terminal
area. This appeared to fix some issues.