1. Added some tests.
2. Moving the gtk makes to 3.0.

Normalized the names for the X windows verion and the GTK version
of graph.c.

Added gtk demo, yes, hello world, but in Cairo and GTK 3.0.

There are two demos with any new sublayer of Petit_ami that are
interesting to me, because they provide proof that the underlying
implementation will support Petit_ami, the buffered and unbuffered
tests. This is because these are the main modes that Petit_ami works
with.