Bringing the Windows build up to the state of Linux.
1. Force Windows builds to use stdio.c. Windows/mingw uses msvcrt.dll,
Microsoft's own emulator. It allows me to replace stdio functions just
as for an .so, but glibc is not used so we cannot use a modified glibc
(it also does not appear to be possible to build a static gtk application.

2. (somewhat) organized the gtk lib/build info.

Bracketed some linux only header files with a linux define tag. Finding
out how to keep the graph_gtk OS independent is going to be an ongoing
concern.

PS. If you get strange runtime errors from gtk binaries, find if you are
running multiple libstdc++-6.dll file.