Restored old definition of test.
A side note, I looked at the situation for terminal level support on
windows. Pascaline used console mode, which originally was derived from
ansilib but now is quite different. Its still possible to use ansi support
in Windows, but the result would be fairly braindamaged and I don't think
it would be satisfactory. The windows solution is to translate the Windows
console library. Thus for the short term, I plan to work on graph_gtk
and maintain the linux/windows commonality for that.

The good news is that gtk seems happy to cross translate my test example
between the systems.