1. Filled out a few more terminal functions (and others).
2. forwarded plcchr().

Scroll is next, need to get this to work for several situations in
terminal mode. The goal is to get terminal mode to work, then pass
screen test, that is level 1. As you might guess, graph mode is
level 2, then Window mode (level 3) and finally widget mode (level 4).

pa_event() is usually where a lot of work lies.