Added lots of key definitions, we are aiming for the same keys as xterm.c.
One small change, X11 does not hyjack one of the function keys like xterm
does, so we present all 12 without modification.

PS. 2 commits ago I left you hanging on an issue in the code.
In plcchr() there is a serious bug. plcchr() puts a single character on the
screen, and it must update both the buffered screen as well as the active
area onscreen. It should only need to draw the same character on the active
screen and be done with it. However, I tried that, and the characters drawn
don't show up until the next character is placed, its very odd. I tried
XFlush() and XSync(), and even tried to get the output queue counts. It
does not work, and the queue counts say 0, meaning nothing is pending.

The workaround was to repaint the entire buffer, which is very crude and
will have serious performance issues, since that happens on every character
output. Also, I presume this same issue happens with other drawing figures,
such as lines, rectangles, etc.

I tried switching to a larger string with the idea that the issue was
limited to single character outputs, but that had no effect.