These are just some proof of concept programs from around the web.
List ALSA ports.

Test play midi file.

Test play wav file.

A sample FM synth.

The subject of software synths is frankly a mess in Linux. Other
systems like Windows and OS X provide at least a minimal software
synth, so programs can count on it being available, even if the
hardware doesn't include it anymore. Since sound.c was written back
in 2002, things have changed. Software synths got better, with
soundfonts, but (perhaps because of that) most or perhaps all,
hardware sound systems don't include a synth.

Its not hard to plug a software synth into Linux, but its frankly
ridiculous to expect a user of a program to go through the rather
arcane process of hooking it up and addressing it, especially if
the sound output is incidental to the operation of the program.

This is why bolting a software synth to the program is a good idea,
and why I may have to look into adding one.