ALSA has issues with calls from .so files (bug filed). The workaround is to use sound.o instead of including in the .so archive.
Improvements on network, still far from operational. See also note below.

No longer needed.

1. Added pa_ coining.

2. alsaplaywave() removed static type, that is an error, will be fixed
next commit.

3. Sundry fixups.

pa_playwave() now operational. It blocks, so this needs improvement. We
are going to have to execute that code in a thread.

For networking, Linux (and I assume MAC OS X) is special in that it is
possible to use stdio against a socket if you use functions that convert
the socket handle to a FILE* handle. This is due to good design on the
part of the sockets system (Berkeley). This means that the network
module may just add overhead.

So what does the network lib do for you? Well:

1. Encapsulates the method, so that it is portable to any OS (including
and especially Windows, which cannot do this with sockets).

2. Simplifies the opening of a TCP/IP connection.

I realize full well that this may not be enough to justify use of the
module. That's why Petit-ami is modular.