Added test files for dumpmidi. This script finds all .mid or .MID files anywhere in tree and runs dumpmidi over them. Generally all the .mid files are placed in a subdirectory, midi, and this test script finds them and tests with them.
The reason this is important is that dumpmidi uses the same parser as
the midi player in sound.c, so this is a quality test using those files.
There are very large collections of .mid files that can be found on the
internet.

The test results I found were that there were a lot of malformed or
outright bad files running around the internet (one file I found
labled as a .mid file was actually a document file). I matched the
runs of this vs. the player from ALSA and found them comparable, so
I am satisfied with the parser and done with testing for now.