Device names fix. Device names are now enumerated and assigned numbers from the table.
1. Increased possible devices to 100. Alsa actually defines quite a few.

2. Added pcm and midi port name tables.

3. pa_opensynthout() now uses the first name in the table. On my system,
this is the default device, and the hope is that this is the same
everywhere.

4. Fixed port lookups to use p-1 for zero based table.

5. Added readalsadev() routine to fill device tables from ALSA.

6. Fill up the pcm out and midi out device name tables.

Connecting up a software synth to ALSA requires considerable setup
work. I don't think this proper to petit_ami, because I don't want
users to have to execute arcane commands to get normal programs
to work. The way around this is to give petit_ami a loadable syth.
This will be a subject for future research.