1. Changed alsaerror() output format.
2. Added ALSA format code to snddev. This is used to set formats delayed on
output devices. We allow the user to set all the parameters after opening
the output device, then the format is made on the first write attempt.

3. Added routines alsa2params() and (convert ALSA format codes to parameters)
and params2alsa() (convert parameters to ALSA format codes). The existing code
was reformatted to use them.

4. pa_openwaveout() refactored to not do format sets at this time, but rather
set the format flag to -1, not set.

5. Fixed pa_chanwaveout().

6. All output parameter sets nullify the format.

7. pa_wrwave() no sets the format if it is undefined.

8. pa_openwavein() now uses the parameters to set format, instead of using
fixed.

9. findparams() changed to use routines above.

This marks the removal of the kludge code. Getting closer to operational
alpha status.