Added boilerplate for synthesizer direct reads and writes.
1. Added equate for default synth input.

2. Added sequencer message types.

3. Added pa_opensynthin(), pa_closesynthin(), pa_wrsynth() and
pa_rdsynth().

The synthesizer direct calls gives a way to read any synth stream from
an external device. The data is formatted in the same synth records the
sequencer uses, and in fact this can be used to interface to the sequencer.
The output write statements give a way to output those same sequencer
records without decoding them, otherwise the functionality is the same
as the "create a midi message" calls.

Moved the sequencer types to the public interface.