1. Changed PREFRATE to MAXINRATE.
2. Added MAXINCHAN and MAXINBITS definitions.

3. The maximum equates are now used against input (capture) devices only to
set parameters. See below.

4. pa_rdwave() and pa_wrwave() get better error handling. Turns out ALSA
gives errors frequently, and the way to handle them is ignore and continue.

So the story with this is ALSA can be quite ambiguous with channel parameters.
They are trying to be accomodating about rates, formats, etc, even on input
devices. This is great, but it makes it hard to figure out what the ideal rate
is for any input. We have taken the approach that input parameters are what
indicates the actual maximum parameters for a device, and that the job of
outputs is to make that happen. Unfortunately ALSA does not always give us
the native format of an input.

So the way we handle this is to limit input devices to "reasonable" values,
which are 44100hz sample rate, 8 channels, and 32 bit floating point. This
keeps the system running, even if a bit wasteful. An input might have only one
channel, and (say) 16 bit signed samples, but through ALSA pseudo devices
this gets extended to 8 channels (by duplication) and 32 floating point
samples by convertion. Most computers can actually handle this in real time.

The way to get the "real" sample formats of inputs, and thus minimum waste
on real time PCM, is to choose the hard devices in ALSA. These are usually
hw:x,y format devices. Thus you get the "right" data by choosing the correct
alsa device. And here are the consequences of that decision:

1. You might have to choose a particular device in ALSA to get what you want.

2. In my testing, none of that is fatal. Real time connection (copy all input
to output) appears to work fine even with nonsense settings, IE, 8 channel
32 bit float.

3. The program could choose the device, but that is really against PA design
rules, which is why I still don't expose the device names.

4. The "default" I/O devices always appear to end up first on the config list,
which means choosing the PA defaults is the right thing.

5. Choosing the input and output devices for default using the Linux menu
ends up changing the "default" ALSA devices. This is the way it should work.

4. ALSA provides a lot of flexibility in its configuration files (yea!).
So that appears (to me) to be where the ALSA specific channel choosing
and format selection can be left to changing those files.

Summing up:

1. The way the system works now is fine if you are willing to do ALSA
configurations.

2. There is a strong case for exposing ALSA device names. I feel this will
happen.