1. Supression of ALSA init error messages is now a flag.
2. Fixed the behavior of readalsadev() on unable to find parameters
for device.

3. Qualified the free for iot, it was probably ok in any case. free()
most likely ignores a null argument.

4. Systematized table print diagnostics.

There is a consistent problem with ALSA as follows:

When enumerating the devices, alsa throws an "unable to open slave" error,
and thus that device is skipped. This only happens when repeatedly runnning
an enumeration, IE., you run an ALSA enabled program back to back multiple
times.

If you simply then wait a minute between runs, the issue goes away. Very
strange.