1. Qualified finding of PCM parameters to wave devices, not MIDI.
2. Qualified table print diagnostic to not print wave parameters on
midi.

Note: ALSA returns "busy" errors on devices that are in use by the system.
That is, go to the sound menu, select a device, and those device(s) go
busy and consequentially will not show up in our enumeration.

I have no idea what to do for this. We can't use a device we can't get
parameters from.