Added keyboard and playmidi programs.
The definition of boolean was causing problems when appearing in multiple
files, so started a new include with commonly used definitions. This is used
in many places.

1. Changed to use localdefs.h.

2. Changed note, channel and instrument to pa_note, pa_channel and pa_instrument
defines to match the convention. This caused multiple changes across multiple
files.

1. Changed to use localdefs.h

2. Changed timer repeat flag to boolean. Boolean was an import from Pascaline,
but its a good convention, matches the one in C++, and completely backwards
compatible. In short, I like it.

Pretty much pa_ coining as described above.

Timer rerun flag change, as above.

New translation. I have no idea what it is doing with those key down timers.
Need to research that.

The keyboard program is a good program for general sound.c testing, but needs
some other capabilities to fully test the api. I brought it back because
alternatives, like VMPK, while nice, offer only a small subset of midi
functionality.

Got tired of recreating this program. Its useful. Also should do playwave.c
and perhaps a few other simple programs. Connect.c? It would be nice not to
have to use the ALSA alternatives for this.