Added "random" program to sound toys.
Added "callback", excseq(). See later.

Send st_playsynth and st_playwave back to main code to be executed.
This needed to happen since they are not MIDI commands. playsyth is
even sort of a macro operation, since it takes the MIDI command list
that was registered previously and decodes it back to syth commands,
and those end up in fluidsynth again.

Changed excseq to global to allow callback.

Added random music generator. This was one of my favorite test tools,
and now is a command (a useless one).

This pretty much closes out the sound.c module. Wow, that was a ride.
A remaining issue would be that the sound from fluidsynth seems to start
scratchy everytime, but fixes itself after repeated use. Very strange.
It might be worthwile to go after some of the realtime concerns that
fluidsynth throws as warnings (suppressed by sound.c).

In any cases such a fix would violate the "out of the box" principle
of Petit_ami (it should work when the program is run, and not require
the user to make system modifications). This is why Fluidsynth is
a plug-in in the first place (to prevent the user from needing to
fiddle with installing and configuring an ALSA plug-in synth).