Turned down the sound by half, to match my spotify at full volume :-).
1. Moved the table print diagnostics up.

2. Refactor plug in adds to add sequentially at the top of the table,
ie., if you add plug1 and plug2, these end up as plug1=1 then plug2=2
at table top. This was needed to make plug in numbering predictable.

The assumption is that you want plug-ins to take over the default
devices. It should probably be an option, the other one being to add
plug-ins at the last table entry.

We don't really have a story on keeping the PA_SYNTH_EXT number
relevant. The idea of that was to be able to select the external output
on MIDI (which indeed would apply to input as well). Honestly I have
been putting that off until I get a Roland MIDI to USB adapter and
seeing how that behaves.