Fixed problem where ssl files would loop on same file descriptor.
1. Added new error.

2. Added new header.

3. Added shadow fid for secured files. Note we may go back to having
opnfil[] be a pointer array and allocate on demand due to the size
of the record.

4. Fixed all of the invalid handle checks. 0 is a valid handle!

5. Create shadow file id entry on secured files.

6. Link the ssl for secured files to the shadow descriptor. More below.

7. Added close for the shadow entry to iclose(). Not real sure if this
is necessary or valid.

So, you are of course sitting and wondering what the daylights is a
shadow file descriptor and why we needed it? We can redirect reads and
writes to their OpenSSL equivalents, but recall! OpenSLL itself is
layered on the file system. That means iread() calls SSL_read(), but
that calls iread again, etc.

We dodge this by duplicating the file descriptor into a non-secured
entry that shadows the original file descriptor. SLL_read()/SLL_write
uses the shadow entry to call back to read/write, which then goes
to the overload routines iread() and iwrite(). Since the shadow is
not marked secured, it passes through to the OS as a plain operation.
No looping.

Anyways, got a test running with downloading a sample html page via
https. Very cool.