Adding multithread protection to network module. This is the first cut after successful compilation.
Changes to be committed:
      modified:   linux/network.c

1. Added per file lock to file record.

2. Added next entry to file record.

3. Added lock for files array.

4. Added free file entries list.

5. Added lock for free file entries.

6. Added getfil() and putfil() routines that get and put file entries
outside of the files array. The advantage of doing that is that it
isolates the malloc(), which has locking issues, and only needs to
lock the free files list. See more below.

7. Split old getfil() into makfil() and newfil(), which just validates
that a file entry exists (makfil) or makes or reuses a new file entry
(newfil).

8. pa_opennet() refactored for locking.

9. iopen uses makfil() now.

10. iclose refactored to use locking. We do a trick here, we lock the
file entry, make a copy, then invalidate all of the contents, then
unlock the entry. This means we own all of the structures like the
ssl data, certificates, etc, outside the context of the file entry, but
the original appears to be atomically closed. This allows us to call
the ssl layer without locking.

11. iread(), iwrite() and ilseek() refactored to use makfil().

12. Init and deinit the global file table and free list locks.

A brief overview of the MT locking phylosphy

The fids (file ids) are unique and allocated by open(), socket() or dup().
This means they are owned by network module and by the caller until closed.
If the calling module mixes open (or its variants) and close operations
on the same file improperly, it is clearly an error, and what is
important is that we do not crash, but instead indicate a proper error.

The case I wanted to optimize is makfil(), which is called each time a file
is accessed, to validate the file. To quickly test the file is usable,
we just test for file entry defined (allocated), then attempt to allocate
that and place it if not. We could have done the check and allocate atomically,
but I wanted the common case of "file entry exists" to go as quickly as
possible. The key fact here is that some files may be first allocated by
a read or write, since they were opened by the OS (fids 0, 1, and 2).

The import of that is that we can find a file table entry empty, but then
find it allocated after we get a file entry for it via getfil(). This is
why the free files list exists, since we don't otherwise need it. We just
put the entry on the free list, allocated with individual locks and all.
It then can be reused for the next try.

Anyways, I see there is more opportunity for error checking, and of couse
extensive error checking is required.

All of this work should allow the creation of, say, a telnet type program
that uses two threads, one for the client to server writes, and another
for the server to client reads, so that a true terminal emulation is
possible.