Added options to the msgserver test program
=== Files changed ===

1. Add options to msgserver build.

1. Fix minor clerical error.

1. Added options.

2. Parse port from command line.

=== Bug numbers affected ===

=== Notes ===

msgclient and msgserver tested together in both plain (udp) and secure
(DTLS) modes.

DTLS is a very important protocol going forward. The next steps are v6,
which is going to be a bit of a challenge to test. Then we need to put
together the certificate handling, then we reach the end of the network
module.