Documentation updates (always); hooker test.
 Changes to be committed:
       modified:   doc/petit_ami.docx
       new file:   libc/hooktest.c

1. Hooktest is a simplified test of the hooker system.

Windows does not use the (incredibly complex, IMHO) system of hooks
in Linux glibc calls, but rather simple overrides via stdio.c.

The hooker demo shows that this is possible in a program stackup.
It also makes a good demo program, or a template for a hooker
based program.

The stdio.c system came a long way since the first demo, which was
done on both Windows and Linux, of stdio hooking, and has been
generally bulletproofed since that time.