Versioned patch system; more install targets; version dependent network.
=== Files changed ===

	modified:   bin/build_glibc

1. Use versioned patching.

	modified:   bin/create_glibc_override_patch

1. Create versioned patch files.

	modified:   bin/glibc/libc-ovr-2.27.a
	modified:   bin/glibc/libc-ovr-2.27.so

1. New 2.27 libs.

	modified:   bin/install

1. New install targets.

	modified:   bin/libc.so.6
	new file:   bin/ver_glibc

1. This command just prints the current glibc version.

	new file:   glibc/glibc_2.27_override.patch
	new file:   glibc/glibc_2.28_override.patch
	new file:   glibc/glibc_2.29_override.patch
	new file:   glibc/libc_2.27_.map.patch
	new file:   glibc/libc_2.28_.map.patch

1. These are all versioned patch files.

	modified:   linux/network.c

1. The "nocancel" i/o calls were tied to a define flag.
The mechanisim to set this flag will be worked out later.

=== Bug numbers affected ===

=== Notes ===

glibc 2.28 and beyond added new files that must be patched. There
were various ways to handle that, but I decided that the most
secure way was to version the patch files. In that way, an unknown
version could have its patch files created in a supervised way,
from scratch, from editing a previous patch, or coped from an
identical patch. These patch files are fairly short in any case.

We probably need to store the build runs, since they are getting
to long to read from the console.