Fix bug in xterm; zero base arrays in Windows console.
 Changes to be committed:
       modified:   doc/petit_ami.docx

1. Converted the terminal chapter to ANSI C.

       modified:   linux/xterm.c

1. Fixed missed screens access issue.

       modified:   windows/console.c

1. Compensated the tabs and screens array accesses for zero base
due to convertion to ANSI C. xterm was used as a model, and in fact
that is how the above bug was found.

Notes:

1. Console is running the first few pages of scntst.

2. I don't have a plan yet to handle buffer to screen differences
in windows. Right now, if you declare a big scrollback buffer, that
shows in the maxx() and maxy() measurements. xterm does not account
for the scrollback space in the window, so that is the model I want
to emulate.

3. I pulled back the resize event, which was actually implemented
in xterm, because it has logical problems, like if the maxx() and
maxy() dimensions dynamically change. That violates the terminal
spec. I don't have an answer at this time.