Improve speed; Add program banner; fix bug in up().
 Changes to be committed:
       modified:   tests/scntst.c

1. Minor speed improvements in character speed test. It didn't help much,
see below.

       modified:   tests/term.c

1, Added signon banner. Otherwise hard to know program started since
windows load of new programs is ssssssssssllllllloooooooooowwwwwww...

       modified:   windows/console.c

1. Decrements in iup() were wrong.

Notes:

1. Pass scntst COMPLETE.

2. Need fix for buffer vs. screen sizes.

3. I performed a speed test on the command line for character output
and got the old numbers, so clearly the time sink is in the program.
It makes no sense, because we are using the same calls as before.

Should try another speed study using just the console output calls
to get the minimum possible speed.