Removed boolean types; removed alsa imported types; added stub files.
 Changes to be committed:
       modified:   include/network.h

1. Changed boolean types to int. boolean is commonly defined in C, which
unfortunately means that it conflicts with any type that we define (hint:
it should be an enum). Its also not in common use at ANSI C, although it
is popular later, and we are aiming for the minimum standard. Its always
possible to define BOOL as int, which is common.

       modified:   include/sound.h

1. Changed boolean types to int (as above).

2. Removed ALSA specific types. This makes the file non-portable. I realize
it creates problems back in Linux, but this needs some other kind of fix.

       new file:   stub/graph.c
       new file:   stub/network.c
       new file:   stub/services.c
       new file:   stub/sound.c
       new file:   stub/terminal.c

3. Defined stub files for all the base Petit-ami files. This will be used
to reform the make system to work on all OSes, even if they don't yet
implement everything.

Notes:

1. The next step is to unify the make file so that we can make under both
Linux and Windows.